feat(comments): refactor comment handling to common views and templates for role-agnostic access
