+++ src/templates/space_admin/issue_management/issue_detail.html
@@ -316,7 +316,7 @@
                class="comments-list-scrollable flex-grow-1">
             
             <!-- Comments will be loaded here via HTMX -->
-            <div hx-get="{% url 'issue_management:space_admin:comment_list' issue_slug=issue.slug %}"
+            <div hx-get="{% url 'issue_management:comment_list' issue_slug=issue.slug %}"
                  hx-trigger="load"
                  hx-swap="innerHTML">
               <div class="text-center text-muted py-4">
@@ -335,7 +335,7 @@
               <span>Comments are disabled for {{ issue.get_status_display|lower }} issues</span>
             </div>
             {% else %}
-            <form hx-post="{% url 'issue_management:space_admin:comment_create' issue_slug=issue.slug %}"
+            <form hx-post="{% url 'issue_management:comment_create' issue_slug=issue.slug %}"
                   hx-target="#comments-list-area"
                   hx-swap="innerHTML"
                   class="comment-form">
