--- src/templates/space_admin/issue_management/partials/comment_list.html
@@ -1,33 +1,0 @@
-<!-- Comment List Only (Form is now separate and sticky) -->
-<div id="comments-list">
-  {% for comment in comments %}
-  <div class="comment-item border-bottom pb-3 mb-3">
-    <div class="d-flex align-items-start flex-column flex-sm-row">
-      <div class="comment-avatar me-0 me-sm-3 mb-2 mb-sm-0 align-self-start">
-        <div class="bg-primary text-white rounded-circle d-flex align-items-center justify-content-center" 
-             style="width: 40px; height: 40px; font-size: 14px; font-weight: bold;">
-          {{ comment.user.get_full_name|default:comment.user.email|slice:":2"|upper }}
-        </div>
-      </div>
-      <div class="comment-content flex-grow-1 w-100">
-        <div class="comment-header d-flex flex-column flex-sm-row align-items-start align-items-sm-center mb-2">
-          <strong class="comment-author me-0 me-sm-2 mb-1 mb-sm-0">
-            {{ comment.user.get_full_name|default:comment.user.email }}
-          </strong>
-          <small class="text-muted">
-            {{ comment.created_at|timesince }} ago
-          </small>
-        </div>
-        <div class="comment-text">
-          {{ comment.comment|linebreaks }}
-        </div>
-      </div>
-    </div>
-  </div>
-  {% empty %}
-  <div class="text-center text-muted py-4 comments-empty">
-    <i class="fas fa-comments fa-2x mb-3 opacity-50"></i>
-    <p class="mb-0">No comments yet. Be the first to comment!</p>
-  </div>
-  {% endfor %}
-</div>
